requirements.txt
agno
openai
python-dotenv
duckduckgo-search
pypdf
lancedb
streamlit
yfinance
pandas
tantivy